Delete count && fix deletion

+++ src/org/horrgs/tempdeleter/TempWindow.java
@@ -11,6 +11,7 @@ import java.io.File;
  */
 public class TempWindow extends JFrame {
     private JButton delete;
+    private JTextArea deleted;
     public TempWindow() {
         setSize(300, 100);
         setTitle("TempDeleter v1.0 by Horrgs");
@@ -18,8 +19,14 @@ public class TempWindow extends JFrame {
         GridBagConstraints gbc = new GridBagConstraints();
         delete = new JButton("Delete");
         gbc.gridx = 0;
-        gbc.gridy = 0;
+        gbc.gridy = 1;
         add(delete, gbc);
+        deleted = new JTextArea("Waiting for files to be deleted.....");
+        deleted.setBackground(getBackground());
+        deleted.setEditable(false);
+        gbc.gridx = 0;
+        gbc.gridy = 0;
+        add(deleted, gbc);
         delete.addActionListener(new TempActionListener());
         setVisible(true);
     }
@@ -31,9 +38,24 @@ public class TempWindow extends JFrame {
             if(ev.getSource() == delete) {
                 String sep = System.getProperty("file.separator");
                 String us = System.getProperty("user.name");
-                File dir = new File("C:" + sep + "Users" + sep + us + sep + "AppData" + sep + "Local" + "Temp" + sep);
-                Util u = new Util();
-                u.deleteDir(dir);
+                File dir = new File("C:" + sep + "Users" + sep + us + sep + "AppData" + sep + "Local" + sep + "Temp" + sep);
+                if(dir.isDirectory()) {
+                    int a = dir.list().length;
+                    deleted.setVisible(true);
+                    Util u = new Util();
+                    u.deleteDir(dir);
+                    int b = dir.list().length;
+                    int amount = a - b;
+                    if(amount == 0) {
+                        deleted.setText("You deleted no files!");
+                    } else if(amount == 1) {
+                        deleted.setText("You deleted " + amount + " file!");
+                    } else if(amount > 1) {
+                        deleted.setText("You deleted " + amount + " files!");
+                    } else {
+                        deleted.setText("What did you delete?!");
+                    }
+                }
             }
         }
     }
